Created a styles.css file for my CSS
Added my CSS styling to the code. Nothing over the top, just to give the program a "look". Lots and lots of room for refining and improving the styling, that is definite. However, I am not looking to do that now but perhaps later because the project was mainly focused on the React.js code.